mynfs open & lseek updated mynfs write/read fixed

+++ mynfslib/include/mynfslib.h
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-int16_t mynfs_open(char const *host, char const *path, uint8_t oflag);
+int16_t mynfs_open(char const *host, char const *path, uint16_t oflag);
 int16_t mynfs_read(char const *host, int16_t fd, void *buf, int16_t count);
 int16_t mynfs_write(char const *host, int16_t fd, void *buf, int16_t count);
 int32_t mynfs_lseek(char const *host, int16_t fd, int32_t offset, uint8_t whence);
